% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ris.R
\name{read_ris}
\alias{read_ris}
\title{A Function to Read in a .ris File of Academic Papers}
\usage{
read_ris(file, ...)
}
\arguments{
\item{file}{A character string of the path to the .ris file to read in.}

\item{...}{Additional arguements to pass to readLines.}
}
\value{
3 tidy data frames: papers, authors, authors_on_papers. Each paper has a unique id (paper_id),
as does each paper (paper_id). Authors and papers are linked by authors_on_papers which links
paper_id, and author_id.
}
\description{
This function reads in a .ris file of academic papers and then
munges the output into 3 tidy data interlinked data frames.
}
\examples{

}
