% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ris_internal.R
\name{read_ris_internal}
\alias{read_ris_internal}
\title{An Internal Function to Read in .RIS Files}
\usage{
read_ris_internal(file, ...)
}
\arguments{
\item{file}{A character string of the path to the .ris file to read in.}

\item{...}{Additional arguements to pass to readLines.}
}
\value{
A nested list for each reference in the .ris file.
}
\description{
A function to read in generic .RIS  files, this function was originally
part of the \href{https://github.com/cran/ris}{ris r package} written by Stephanie Kovalchik.
}
\examples{

}
